platform/x86: ISST: Check validity of instance before use
In some unlikely scnerio when the TPMI SST data is invalid in some of
the packages, there will be no TPMI SST instance allocated for those
packages.

But there is no check for validity of instance in the function:
isst_if_get_tpmi_instance_count()

Add a check for validity of the instance pointer before trying to
dereference the pointer.

Signed-off-by: Srinivas Pandruvada <srinivas.pandruvada@linux.intel.com>