x86/microcode/intel: Check update requirements on CPUs supporting rollback
With new refactoring we are checking for pending commit right at the
entry of `request_microcode_fw`. This check is only valid if
commit_svn == 1. Since we don't even proceed if commit_svn == 1,
we can safely remove the pending_svn check.

This commit also fixes the reload failure seen on legacy and GNR
platforms.

Suggested-by: Ashok Raj <ashok.raj@intel.com>
Signed-off-by: Vijay Dhanraj <vijay.dhanraj@intel.com>