tsc: use the better socket number estimation
This is temp workaround patch for tsc issue on GNR with SNC3 mode
enabled, which will use a better way to estimate the number of
sockets in system.

The code is mostly from upstream kernel maintainer Thomas Gleixner,
while combined with some code from Intel team. More details are in
the following link:

  https://lore.kernel.org/lkml/87edm36qqb.ffs@tglx/
  https://lore.kernel.org/lkml/20230613052523.1106821-2-feng.tang@intel.com/

Also the socket check limit is lifted to 4, as some customer met
this before, and the lift patch has been in linux-next tree for
a while.

Signed-off-by: Lijun Pan <lijun.pan@intel.com>