x86/microcode/intel: Fix "BUG: sleeping function called from invalid context"
When doing the stress test of CPUs online/offline, the following
calltrace occurred:

   ? ___slab_alloc+0x424/0x780
    __might_resched.cold+0xd6/0x10f
    __kmem_cache_alloc_node+0x3d7/0x4d0
    ? uc_doe_create_mbox+0x3e/0xa0
    kmalloc_trace+0x29/0xa0
    uc_doe_create_mbox+0x3e/0xa0
    ? completion_done+0x2e/0x40
    ? __pfx_cache_ap_init+0x10/0x10
    ? stop_machine_from_inactive_cpu+0x104/0x140
    ? __pfx_cache_rendezvous_handler+0x10/0x10
    collect_cpu_info+0x188/0x1f0
    ? __pfx_mc_cpu_starting+0x10/0x10
    mc_cpu_starting+0x4b/0xb0
    cpuhp_invoke_callback+0x249/0x4f0
    __cpuhp_invoke_callback_range+0x79/0xc0

The microcode mailbox instances are created/destroyed during CPUs
online/offline, which occurs in an atomic context. The current
microcode mailbox driver allocates memory for the microcode instances
with the "GFP_KERNEL" flag that, leading to the calltrace. Fix this
issue by using the "GFP_ATOMIC" flag when allocating memory for the
microcode instances.

Fixes: 5a399abd7675 ("x86/microcode/intel: Add microcode DOE mailbox driver")
Reported-by: Wang Wendy <wendy.wang@intel.com>
Signed-off-by: Qiuxu Zhuo <qiuxu.zhuo@intel.com>