tboot: Disable LASS before calling tboot shutdown
On SRF we'd get a GPF when calling tboot because we're in the upper
half of the address space while the tboot code, being 1:1 mapped, is
quite likely in the lower half.  LASS protects from this type of calls.

We probably want all types of memory protection features off since
tboot didn't enable them and may have been compiled without
support for them.  Ideally we want to return control to tboot with
the CPU in the same state as it was when tboot started Linux.

A simple stac() doesn't prevent the GPF in testing.

Signed-off-by: Andrew Zaborowski <andrew.zaborowski@intel.com>