cxl/region: Calculate an endpoint position in a region interleave
Regions must be assembled in correct target interleave order,
and there is only one correct order.

For a user defined region, the user needs to out the targets.
to the user the targets are memdevs, but in driver language
the targets are actually the endpoint decoders assigned to
decode an hpa range in that memdev.

This calculation offers a method to determine and endpoints
position in the region interleave by climbing up it's ancestral
tree to the root decoder.

The position calculation start with the endpoints local position,
ie it's position in it's parents port. Traversing up towards
the root decoder and examining position and ways allows the position
location to be refined all the way to the root decoder.

The calculation depends on this a calculation applied iteratively:
pos = pos * parent_ways + parent_pos;

And then there are guardrails when looking for that parent_pos:
Decoder target lists can repeat entrys, so you can have a port
with:   0: portA
	1: portB
	2: portA
	3: portB
In order to match to the correct position in this example [0-3]
an index is created using the child position and ways. So,
here there is a target list of 4, but only 2 children.

Discovering the position in the parent port requires

the position in the region interleave is refined
until the final position is calculated.

The usage of the new algorithm provides an opportunity to sanity
check the drivers calculation against any successful user created
region.  For example, a user requested region gets created, and
upon success, this calculation is run and complains if
check

Signed-off-by: Alison Schofield <alison.schofield@intel.com>
Signed-off-by: Li Ming <ming4.li@intel.com>