cxl/region: Update the region target position sort
Previously, sorting the target list of an auto created region was
done using a relative sorting algorithm. It

Replace that sorting algorithm with the cxl_calc_interleave() that
finds the exact position in an interleave set based on the endpoints
ancestor positions and ways and the region interleave ways. Once
that exact position is known, sort the list into pos order.

cxl_calc_interleave() was introduced in a prior patch, so here
it is only put to use in the cxl_region_sort_targets(), replacing
the previous algorithm, using sort().

Signed-off-by: Alison Schofield <alison.schofield@intel.com>
Signed-off-by: Li Ming <ming4.li@intel.com>